% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_na_between_non_na.R
\name{get_na_between_non_na}
\alias{get_na_between_non_na}
\title{Check if a vector of values has an NA in between at least two non-NA
values. If so, it will return a vector of indices of NA values for which
this is the case. If none are found, it will return an empty vector.}
\usage{
get_na_between_non_na(x)
}
\arguments{
\item{x}{vector of samples}
}
\value{
vector of indices (numeric)
}
\description{
This can e.g. conveniently be used to test whether there are any BLQ
samples in-between two numeric (above LOQ) values. Within a single PK
curve this is not likely to occur and the individual or data point should be
flagged as an outlier.
}
