% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grapes-less-than-equals-grapes.R
\name{\%<=\%}
\alias{\%<=\%}
\title{Less-than-or-equal-to with a little room for floating point precision
issues}
\usage{
x \%<=\% y
}
\arguments{
\item{x}{Numeric vector}

\item{y}{Numeric vector}
}
\description{
Less-than-or-equal-to with a little room for floating point precision
issues
}
\keyword{internal}
