% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irxutils-package.R
\docType{package}
\name{irxutils-package}
\alias{irxutils}
\alias{irxutils-package}
\title{irxutils: Generic utils}
\description{
Small generic functions for e.g. string and object manipulations, datetime parsing, etc.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://insightrx.github.io/irxutils/}
}

}
\author{
\strong{Maintainer}: Ron Keizer \email{ron@insight-rx.com}

Other contributors:
\itemize{
  \item Michael McCarthy \email{michael.mccarthy@insight-rx.com} [contributor]
  \item InsightRX [copyright holder]
}

}
\keyword{internal}
