% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_string.R
\name{random_string}
\alias{random_string}
\title{Random string}
\usage{
random_string(n = 5)
}
\arguments{
\item{n}{number of characters for random string}
}
\value{
A string of random letters with \code{n} characters.
}
\description{
Random string
}
\examples{
random_string(3)

}
