% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grapes-greater-than-equals-grapes.R
\name{\%>=\%}
\alias{\%>=\%}
\alias{\%<=\%}
\title{Safe relational operators}
\usage{
x \%>=\% y

x \%<=\% y
}
\arguments{
\item{x, y}{Numeric vectors.}
}
\value{
A logical vector indicating the result of the element by element comparison.
}
\description{
Binary operators which allow the comparison of values in atomic vectors, with
a little room for floating point precision issues.
}
\details{
These binary comparison operators make the base relational operators
(\link[base:Comparison]{base::Comparison}) safer by adding a little room for floating point precision
issues. \verb{\%>=\%} is the counterpart to \code{>=}, and \verb{\%<=\%} is the counterpart to
\code{<=}.
}
\examples{
(0.7 - 0.4) >= 0.3
(0.7 - 0.4) \%>=\% 0.3

}
