% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_continuous.R
\name{is_continuous}
\alias{is_continuous}
\title{Test if a vector of values is likely to be continuous or categorical}
\usage{
is_continuous(x, cutoff = 0.8)
}
\arguments{
\item{x}{vector of values}

\item{cutoff}{cutoff value for deciding between continuous/categorical}
}
\description{
The function will try to convert all values to numeric. If a certain fraction
of values is able to convert succesfully, it will assume the vector is
continuous. Default treshold is 0.8.
}
